requirements.txt
openai~=1.60.2
python-dotenv~=1.0.1
requests~=2.32.3
fastapi~=0.115.8
uvicorn~=0.34.0
caldav~=1.4.0